added module test
