--- src/Select.jl
@@ -1,7 +1,0 @@
-module Select
-
-    include("select/_select.jl")
-    include("select/buffers.jl")
-    include("interpolations.jl")
-
-end # module
