+++ src/graphics/_graphics.jl
@@ -1,7 +1,8 @@
 using Plots, LaTeXStrings
 using PyPlot: streamplot, imshow, figure
 import Makie
-using JuliaDispatch.Utils, JuliaDispatch.Buffers, JuliaDispatch.Select
+using JuliaDispatch.Utils, JuliaDispatch.Buffers
+using JuliaDispatch.Select: patches_in
 gr()
 default(:size, (1200, 800))
